requirements.txt
asgiref==3.9.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.7.0
djangorestframework==3.14.0
PyMySQL==1.1.0
filelock==3.18.0
fsspec==2025.5.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper==20240930
Pillow==10.1.0
python-decouple==3.8
regex==2024.11.6
requests==2.32.3
setuptools==80.8.0
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.6.0
gunicorn==21.2.0
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
social-auth-app-django==5.4.0
pillow==10.1.0
django-storages==1.14.2
channels==4.0.0
channels-redis==4.1.0
